Make `-F winapp` buildable on all platforms
It builds as if `-F winapp` is not specified on other platforms than Windows.
