+++ Cargo.toml
@@ -23,6 +23,8 @@ regex = "1.11.1"
 simplelog = "0.12.2"
 strum = "0.27.1"
 strum_macros = "0.27.1"
+
+[target.'cfg(windows)'.dependencies]
 toast-logger-win = { version = "0.5.0", optional = true }
 
 [features]
@@ -38,4 +40,4 @@ path = "src/main.rs"
 [[bin]]
 name = "monitor-inputw"
 required-features = ["winapp"]
-path = "src/main_winapp.rs"
+path = "src/main.rs"
